chatbot.py
import streamlit as st
import langchain
from langchain.llms import OpenAI
from langchain.chains import RetrievalQAWithSourcesChain
from langchain.document_loaders import UnstructuredURLLoader
from langchain_community.vectorstores import FAISS
from sentence_transformers import SentenceTransformer
import faiss
import os

# Setup (Adjust according to your initial setup and imports)

llm = OpenAI(temperature=0.9, max_tokens=500)
user_question = st.text_input("Enter your question:")
if st.button('Answer Question'):
    # Your processing code here
    # For simplicity, let's assume you've set up everything needed for LangChain
    # And you have a function `get_answer(question)` that utilizes your LangChain setup

    answer = get_answer(user_question)  # Simplified example, replace with your actual function call

    st.write(answer)
